Don't use a 32-bit bit type as offset argument for kvm dirty tracking
In the call path of kvm_get_dirty_pages_log_range(),
its caller kvm_get_dirty_bitmap_cb() passes the
target_phys_addr_t both as start_addr and the offset.
So, using int will make dirty tracking over 4G fail
completely.

+++ qemu-kvm.c
@@ -1190,7 +1190,7 @@ int kvm_physical_memory_set_dirty_tracking(int enable)
 /* get kvm's dirty pages bitmap and update qemu's */
 static int kvm_get_dirty_pages_log_range(unsigned long start_addr,
                                          unsigned char *bitmap,
-                                         unsigned int offset,
+                                         unsigned long offset,
                                          unsigned long mem_size)
 {
     unsigned int i, j, n=0;
